+++ packages/storybook/tailwind.config.ts
@@ -2,7 +2,7 @@ import { jamsrui } from "@jamsr-ui/theme";
 import type { Config } from "tailwindcss";
 
 const config: Config = {
-  content: ["../components/**/*.@(js|jsx|mjs|ts|tsx)"],
+  content: ["./node_modules/@jamsr-ui/**/dist/*.{js,jsx,mjs}"],
   plugins: [...jamsrui()],
 };