+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/menu
 
+## 0.0.17
+
+### Patch Changes
+
+- added dts
+
 ## 0.0.16
 
 ### Patch Changes