+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ui/storybook
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/react@0.0.21
+
 ## 0.0.21
 
 ### Patch Changes